requirements.txt
pyinstaller
moviepy
python-vlc
pandas
csv
pandas
toml
